move more oc leaf packages to staging
