+++ staging/src/github.com/openshift/oc/pkg/cli/admin/release/extract_tools_test.go
