+++ staging/src/github.com/openshift/oc/pkg/cli/admin/release/image_mapper.go
