+++ staging/src/github.com/openshift/oc/pkg/cli/admin/release/info_test.go
