+++ staging/src/github.com/openshift/oc/pkg/cli/admin/release/mirror.go
@@ -26,9 +26,9 @@ import (
 	imageclient "github.com/openshift/client-go/image/clientset/versioned"
 	"github.com/openshift/library-go/pkg/image/dockerv1client"
 	imagereference "github.com/openshift/library-go/pkg/image/reference"
-	"github.com/openshift/origin/pkg/oc/cli/image/extract"
-	imagemanifest "github.com/openshift/origin/pkg/oc/cli/image/manifest"
-	"github.com/openshift/origin/pkg/oc/cli/image/mirror"
+	"github.com/openshift/oc/pkg/cli/image/extract"
+	imagemanifest "github.com/openshift/oc/pkg/cli/image/manifest"
+	"github.com/openshift/oc/pkg/cli/image/mirror"
 )
 
 // NewMirrorOptions creates the options for mirroring a release.
