+++ staging/src/github.com/openshift/oc/pkg/cli/admin/release/release.go
