+++ staging/src/github.com/openshift/oc/pkg/cli/admin/release/test/main.go
