+++ staging/src/github.com/openshift/oc/pkg/cli/idle/idle.go
