+++ staging/src/github.com/openshift/oc/pkg/cli/image/archive/archive.go
