+++ staging/src/github.com/openshift/oc/pkg/cli/image/archive/archive_linux.go
