+++ staging/src/github.com/openshift/oc/pkg/cli/image/archive/time_linux.go
