+++ staging/src/github.com/openshift/oc/pkg/cli/image/extract/extract.go
@@ -27,9 +27,9 @@ import (
 	"github.com/openshift/library-go/pkg/image/dockerv1client"
 	imagereference "github.com/openshift/library-go/pkg/image/reference"
 	"github.com/openshift/library-go/pkg/image/registryclient"
-	"github.com/openshift/origin/pkg/oc/cli/image/archive"
-	imagemanifest "github.com/openshift/origin/pkg/oc/cli/image/manifest"
-	"github.com/openshift/origin/pkg/oc/cli/image/workqueue"
+	"github.com/openshift/oc/pkg/cli/image/archive"
+	imagemanifest "github.com/openshift/oc/pkg/cli/image/manifest"
+	"github.com/openshift/oc/pkg/cli/image/workqueue"
 )
 
 var (
