+++ staging/src/github.com/openshift/oc/pkg/cli/image/info/info.go
@@ -27,8 +27,8 @@ import (
 	"github.com/openshift/library-go/pkg/image/dockerv1client"
 	imagereference "github.com/openshift/library-go/pkg/image/reference"
 	"github.com/openshift/library-go/pkg/image/registryclient"
-	imagemanifest "github.com/openshift/origin/pkg/oc/cli/image/manifest"
-	"github.com/openshift/origin/pkg/oc/cli/image/workqueue"
+	imagemanifest "github.com/openshift/oc/pkg/cli/image/manifest"
+	"github.com/openshift/oc/pkg/cli/image/workqueue"
 )
 
 func NewInfoOptions(streams genericclioptions.IOStreams) *InfoOptions {
