+++ staging/src/github.com/openshift/oc/pkg/cli/image/manifest/dockercredentials/credentials_test.go
