+++ staging/src/github.com/openshift/oc/pkg/cli/image/manifest/errors.go
