+++ staging/src/github.com/openshift/oc/pkg/cli/image/manifest/manifest.go
@@ -26,8 +26,8 @@ import (
 	"github.com/openshift/library-go/pkg/image/dockerv1client"
 	imagereference "github.com/openshift/library-go/pkg/image/reference"
 	"github.com/openshift/library-go/pkg/image/registryclient"
+	"github.com/openshift/oc/pkg/cli/image/manifest/dockercredentials"
 	"github.com/openshift/oc/pkg/helpers/image/dockerlayer/add"
-	"github.com/openshift/origin/pkg/oc/cli/image/manifest/dockercredentials"
 )
 
 type ParallelOptions struct {
