+++ staging/src/github.com/openshift/oc/pkg/cli/image/mirror/mappings.go
