+++ staging/src/github.com/openshift/oc/pkg/cli/image/mirror/plan.go
