+++ staging/src/github.com/openshift/oc/pkg/cli/image/mirror/s3.go
