+++ staging/src/github.com/openshift/oc/pkg/cli/registry/info/info.go
