+++ staging/src/github.com/openshift/oc/pkg/cli/registry/login/login.go
